// gestimowebfront-end/src/gs-api/src/models/immeuble-etage-dto.ts
/* tslint:disable */
export interface ImmeubleEtageDto {
  id?: number;
  idAgence?: number;
  idCreateur?: number;
  codeNomAbrvImmeuble?: string;
  nomCompletImmeuble?: string;
  nomBaptiserImmeuble?: string;
  descriptionImmeuble?: string;
  numImmeuble?: number;
  nbrEtage?: number;
  nbrePiecesDansImmeuble?: number;
  idSite?: number;
  idUtilisateur?: number;
  nomPropio?: string;
  prenomProprio?: string;
  garrage?: boolean;
}